reconst bugs

+++ src/task/reconstructor/reconstructorbase.cpp
@@ -656,8 +656,8 @@ void ReconstructorBase::createMeasurement(reconstruction::Measurement& mm,
     mm.ay_stddev = acc_acc.ay_stddev_;
 }
 
-void ReconstructorBase::createMeasurement(reconstruction::Measurement& mm, 
-                                          unsigned int rec_num)
+void ReconstructorBase::createMeasurement(reconstruction::Measurement& mm,
+                                          unsigned long rec_num)
 {
     auto it = target_reports_.find(rec_num);
     assert(it != target_reports_.end());
